Menu/serializers.py
# rest_framework
from rest_framework import serializers

# models
from .models import Item


class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = '__all__'
